+ YearData, MonthType 클래스 삭제; 한 해의 정보를 int 안에 넣어 비트연산. + scrap 테스트코드에서 음력기준데이터 복붙용 코드 바로 생성 + 자바 버전 8

+++ src/test/java/xyz/arinmandri/kasiapi/Item.java
@@ -26,7 +26,7 @@ public class Item
 	@Element( name = "solDay" , required = false ) int solDay;// 양력 일
 	@Element( name = "solWeek" , required = false ) String solWeek;// 요일
 
-	@Element( name = "solJd" , required = false ) long solJd;// 율리우스적일
+	@Element( name = "solJd" , required = false ) int solJd;// 율리우스적일
 
 	public String toLunString () {
 		StringBuilder sb = new StringBuilder()
